+++ lib/libft/ft_atof.c
@@ -6,31 +6,31 @@
 /*   By: igchurru <igchurru@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/01 12:28:39 by igchurru          #+#    #+#             */
-/*   Updated: 2025/04/08 16:48:51 by igchurru         ###   ########.fr       */
+/*   Updated: 2025/05/29 12:19:50 by igchurru         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
 /*
- * Converts a string representation of a floating-point number to a float.
+ * Converts a string representation of a doubleing-point number to a double.
  *
  * The function iterates through the input string, handling an optional
  * leading sign ('-'). It accumulates the integer part of the number
  * and then handles the fractional part after a decimal point ('.').
- * The final float value is calculated by dividing the accumulated number
+ * The final double value is calculated by dividing the accumulated number
  * by the appropriate power of 10 based on the number of digits after
  * the decimal point and applying the sign.
  *
- * @param n A pointer to the null-terminated string representing the float.
- * @return The floating-point number represented by the input string.
+ * @param n A pointer to the null-terminated string representing the double.
+ * @return The doubleing-point number represented by the input string.
  */
 
-float	ft_atof(char *n)
+double	ft_atof(char *n)
 {
 	int		i;
 	int		j;
-	float	nb;
+	double	nb;
 	int		m;
 
 	i = 0;
